move reporting to Bugsnag.Reporter
